# public/openapi.yaml
openapi: 3.0.0
info:
  title: MarlimPay API
  version: v1.0.0
  description: API for MarlimPay, a financial transaction simulator.
servers:
  - url: http://localhost:5001/marlimpay-dev/us-central1/api 
    description: Local Firebase Emulator

components:
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: Unique identifier for the user.
          readOnly: true
        name:
          type: string
          example: Maria Silva
        email:
          type: string
          format: email
          example: maria@example.com
        balance:
          type: number
          format: float
          description: Current balance of the user.
          example: 1000.00
        createdAt:
          type: string
          format: date-time
          description: Timestamp of user creation.
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of last user update.
          readOnly: true

    CreateUserRequest:
      type: object
      required:
        - name
        - email
        - balance
      properties:
        name:
          type: string
          example: Maria Silva
        email:
          type: string
          format: email
          example: maria@example.com
        balance:
          type: number
          format: float
          description: Initial balance for the user.
          example: 1000.00

    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          example: Maria da Silva
        email:
          type: string
          format: email
          example: mariads@example.com

    Transaction:
      type: object
      properties:
        transaction_id:
          type: string
          format: uuid
          description: Unique identifier for the transaction.
          readOnly: true
        payer_id:
          type: string
          format: uuid
          description: User ID of the payer.
        receiver_id:
          type: string
          format: uuid
          description: User ID of the receiver.
        amount:
          type: number
          format: float
          description: Amount of the transaction.
        status:
          type: string
          enum: [pending, approved, failed]
          description: Status of the transaction.
        createdAt:
          type: string
          format: date-time
          description: Timestamp of transaction creation.
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of last transaction update.
          readOnly: true

    CreateTransactionRequest:
      type: object
      required:
        - payer_id
        - receiver_id
        - amount
      properties:
        payer_id:
          type: string
          format: uuid
          example: user_abc123
        receiver_id:
          type: string
          format: uuid
          example: user_def456
        amount:
          type: number
          format: float
          example: 150.75

    CreateTransactionResponse:
      type: object
      properties:
        transaction_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending] 
        created_at:
          type: string
          format: date-time

    UserTransactionItem:
      type: object
      properties:
        transaction_id:
          type: string
          format: uuid
        direction:
          type: string
          enum: [sent, received]
        amount:
          type: number
          format: float
        status:
          type: string
          enum: [pending, approved, failed]

    WebhookPaymentGatewayRequest:
      type: object
      required:
        - transaction_id
        - status
      properties:
        transaction_id:
          type: string
          format: uuid
          example: txn_abc123
        status:
          type: string
          enum: [approved, failed]
          example: approved

    WebhookResponse:
      type: object
      properties:
        message:
          type: string
          example: Webhook received com sucesso

    WebhookLogResponse:
      type: object
      properties:
        payload:
          type: object
          properties:
            transaction_id:
              type: string
              format: uuid
            status:
              type: string
              enum: [approved, failed]
        receivedAt:
          type: object
          properties:
            _seconds:
              type: integer
              example: 1750599721
            _nanoseconds:
              type: integer
              example: 951000000
        processingStatus:
          type: string
          enum: [processed, failed]
          example: processed

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "JWT"

security:
  - BearerAuth: []

paths:
  /api/v1/users:
    post:
      summary: Create a new user
      operationId: createUser
      tags: [Users]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List all users
      operationId: listUsers
      tags: [Users]
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /api/v1/users/{id}:
    get:
      summary: Get user by ID
      operationId: getUserById
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the user to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update user name and email
      operationId: updateUser
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the user to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/users/{user_id}/transactions:
    get:
      summary: List transactions for a user
      operationId: getUserTransactions
      tags: [Transactions]
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID for transactions list
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTransactionItem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/transactions:
    post:
      summary: Create a new transaction
      operationId: createTransaction
      tags: [Transactions]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: UUID for request idempotency
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created (pending)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTransactionResponse'
        '400':
          description: Invalid request or insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404': 
          description: Payer or receiver not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict or Idempotency-Key used
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/transactions/{id}:
    get:
      summary: Get transaction by ID
      operationId: getTransactionById
      tags: [Transactions]
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the transaction
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /webhook/payment-gateway:
    post:
      summary: Webhook for payment gateway
      operationId: paymentGatewayWebhook
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPaymentGatewayRequest'
      responses:
        '200':
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction not found for webhook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /webhook/payment-gateway/logs/{id}:
    get:
      summary: Get webhook log by ID
      operationId: getWebhookLogById
      tags: [Webhooks]
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the webhook log to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Webhook log retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookLogResponse'
        '404':
          description: Webhook log not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'